requirements.txt
google-genai>=0.1.0
requests>=2.28.0 